test(cli): add unit tests for CLI module with mocked external calls
- Add comprehensive unit tests for `main`, `_report_issues`, and `_print`, covering various scenarios.
- Ensure full mocking of external calls for isolated testing.
- Refactor integration tests to focus on subprocess behavior.
